Kaggle Time Series Forecasting Using Linear Regression
Competition in Kaggle: https://www.kaggle.com/competitions/store-sales-time-series-forecasting
It was a real challange to me and it was the first one of many. I haven't done the best code, however it was the best code for me today. I have commented in code all my improves during this challanges, my difficults and what I would like to do if I had more skills. This showed me how much I can learn in ML and Data Analysis and I will work for create a better model for this competition